TP-1445 Update palvelumanuaali theme to use emulsify base theme
Previously, palvelumanuaali theme was based on emulsify_drupal, which is
a starter kit theme. However, it's now minimally maintained. For this
reason palvelumanuaali is now updated to use the new emulsify base theme.

Also, font and image assets are moved to assets directory, which was
created by the theme initialization script.

Some changes are made to webpack configs. For example, sass-loader is
set to use the legacy API to avoid certain errors.