TP-1647 Fix altering access for leaving group
LeaveGroupAccessControl was made to prohibit group admins to leave their
group even though they have the permission to administer group members.
However, the solution did not take into account the different group
relationships. Also, it checked the relationship ownership and not the
actual referenced user. This led to the situation were group admins were
unable to remove any of their own group relations, whether it was
member, content or subgroup. Also, it did not actually prevent group
admins to leave the group if someone else had made the membership
relation.

This fix will narrow the access changes only to the group membership
relations. The referenced user entity is used to check if it matches the
current account user.

This change also fixes TP-1663.