TP-1713 Use role-based access on view to fix missing content
Commit ef6678d2ca33a0dec8f53f069671a5cbdf55bf65 added group permission
to some views to prevent members of other groups to see the group's
content. This partially reverts it by using role-based access again on
a view display that is embedded on a group secondary display. This
allows rendering the group on another view while also using the group
permission on a view display rendered on a group primary display.